[STORYBOOK] Ré-organise les stories du LAB et les déplace dans le dossier legacy
