+++ stories/composants/legacy/Icone.stories.svelte
